Assignments/Personalized_Education_Assistant/requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.4.0
requests>=2.31.0

# AI and ML Libraries
crewai>=0.28.8
langchain-google-genai>=1.0.0
google-generativeai>=0.3.2

# Data Visualization
plotly>=5.17.0

# Additional CrewAI Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Optional but recommended for better performance
typing-extensions>=4.8.0
httpx>=0.25.0